Removed cast for getAIRange that forgot to remove last commit

+++ st_final/source/st_final.cpp
@@ -161,7 +161,7 @@ int stFinal::getDefaultLightSetIndex()
 }
 stRange* stFinal::getAIRange()
 {
-    return (u32) & this->aiRange;
+    return &this->aiRange;
 }
 bool stFinal::isAdventureStage()
 {
