+++ st_targetsmash/source/st_targetsmash.cpp
@@ -167,7 +167,7 @@ int stTargetSmash::getDefaultLightSetIndex()
 }
 stRange* stTargetSmash::getAIRange()
 {
-    return (u32) & this->aiRange;
+    return &this->aiRange;
 }
 bool stTargetSmash::isAdventureStage()
 {
